model collectives data accuracy check
Summary:
A prototype on the flow for offline model collective data checker:

- The model collective traces are replayed both on a reference run, either on CPU or GPU ranks and data checker run is done on Athena ranks.  The reference reduction collective outputs are compared with the collective outputs from Athena (Atol and RTol).  Both Atol and Rtol can be configured.
 ---
 ---
The model comms datachecker flow has 2 steps:
 ---
Step 1. **Generating golden reference collective output:**

Reduction Algorithms (AllReduce and ReduceScatter's input and output are saved in a pickle file locally for each rank and each replay iteration.  In this step, random data is used an input for the collective.
- Golden reference collective output can be from CPU ranks (gloo backend) or GPU (nccl backend).
- A hashID with {collectives name, processgroup_id, ip_msgsize, op_msgsize} is generated and a table entry is created for each collective storing the [ip_tensor, op_tensor] as a golden reference for the data check.
- The hashTable is saved locally.  This hashTable generation can be a one-time step for a model config.

 ---
Step 2: **Collective data check mode mtia with hccl backend:**
The collectives are launched to the device (gpu or mtia).  Using the  the device outputs are checked against the saved collective outputs for the respective rank and iteration.
 ---
 ---

Reviewed By: qcyuan

Differential Revision: D79583482

fbshipit-source-id: 5c40f0bc610876da9341768da2ef70010b1202b5